+++ app/css/index.css
@@ -526,6 +526,10 @@ input[type='password']:autofill {
 	inset: 0px;
 }
 
+.bottom-4 {
+	bottom: 1rem;
+}
+
 .left-4 {
 	left: 1rem;
 }
@@ -548,10 +552,6 @@ input[type='password']:autofill {
 
 .top-2 {
 	top: 0.5rem;
-}
-
-.bottom-4 {
-	bottom: 1rem;
 }
 
 .z-10 {
@@ -962,6 +962,10 @@ input[type='password']:autofill {
 	row-gap: 0.75rem;
 }
 
+.gap-y-2 {
+	row-gap: 0.5rem;
+}
+
 .overflow-hidden {
 	overflow: hidden;
 }
@@ -1031,6 +1035,10 @@ input[type='password']:autofill {
 	border-color: rgb(248 113 113 / 0.2);
 }
 
+.border-red-400\/50 {
+	border-color: rgb(248 113 113 / 0.5);
+}
+
 .border-transparent {
 	border-color: transparent;
 }
@@ -1095,6 +1103,10 @@ input[type='password']:autofill {
 	background-color: rgb(23 23 23 / var(--tw-bg-opacity));
 }
 
+.bg-red-400\/10 {
+	background-color: rgb(248 113 113 / 0.1);
+}
+
 .bg-red-400\/5 {
 	background-color: rgb(248 113 113 / 0.05);
 }
@@ -1126,6 +1138,14 @@ input[type='password']:autofill {
 
 .bg-white\/50 {
 	background-color: rgb(255 255 255 / 0.5);
+}
+
+.bg-black\/80 {
+	background-color: rgb(0 0 0 / 0.8);
+}
+
+.bg-neutral-900\/80 {
+	background-color: rgb(23 23 23 / 0.8);
 }
 
 .p-10 {
@@ -1726,6 +1746,22 @@ input[type='password']:autofill {
 }
 
 @media (min-width: 1024px) {
+	.lg\:right-4 {
+		right: 1rem;
+	}
+
+	.lg\:top-4 {
+		top: 1rem;
+	}
+
+	.lg\:bottom-auto {
+		bottom: auto;
+	}
+
+	.lg\:left-auto {
+		left: auto;
+	}
+
 	.lg\:hidden {
 		display: none;
 	}
